Budget page constructed 50% + Transaction Page improvement
Budget page is now 50% constructed:
- Budget tracker is built, user can specify any number of tags they want to track. And when any expense using the same tag is added, they will accumulated in the budget progress bar. Ex: tracking coffee spending.
NOTE: When user create new expense, they should be reminded of the existing tags, so that they don't use the wrong tag.

-Multiple goals can be added, and each one will have the option to be edited, or deleted. The changes will be reflected immediately. When edit or delete is done, fetching from the database is done again.

Transaction page:
previously, removing and adding new transactions will be updated on the database, but the local transactionlist is re- rendered with a fake entry. --> This causes a bug where new "fake" entries have edit and remove buttons, but when clicked, nothing happens, until page refresh is done.  Now it is done the same way in budget page. Once changes are done, refetching of the database is prompted, and transaction list is re-rendered.